extract OpenEdxDB
